Moved build related code to build.py
The code for building a policy set, based on a cfbs.json, should
be quite straight forward and isolated from all the other code.
